rename dialog folder and move context menu
