+++ src/components/dialogs/NewTabDialog.test.tsx
