+++ src/components/dialogs/NewTabDialog.tsx
